REF: feature_data & evictions
