+++ q2_types/multiplexed_sequences/_transformer.py
@@ -6,7 +6,9 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
+import gzip
 import shutil
+import itertools
 
 import pandas as pd
 import skbio.io
@@ -23,13 +25,19 @@ from . import (
     EMPPairedEndDirFmt,
     EMPPairedEndCasavaDirFmt
 )
-from ..per_sample_sequences import (
-    SingleLanePerSampleSingleEndFastqDirFmt,
-    SingleLanePerSamplePairedEndFastqDirFmt
-)
-from ..feature_data._util import _PlotQualView
 from ..plugin_setup import plugin
 from .._util import FastqGzFormat
+from ._objects import (BarcodePairedSequenceFastqIterator,
+                       BarcodeSequenceFastqIterator)
+
+
+def _read_fastq_seqs(filepath):
+    # This function is adapted from @jairideout's SO post:
+    # http://stackoverflow.com/a/39302117/3424666
+    fh = gzip.open(filepath, 'rt')
+    for seq_header, seq, qual_header, qual in itertools.zip_longest(*[fh] * 4):
+        yield (seq_header.strip(), seq.strip(), qual_header.strip(),
+               qual.strip())
 
 
 @plugin.register_transformer
@@ -79,17 +87,6 @@ def _19(dirfmt: EMPPairedEndCasavaDirFmt) -> EMPPairedEndDirFmt:
     return result
 
 
-# TODO: Remove _PlotQualView once QIIME 2 #220 completed
-@plugin.register_transformer
-def _30(dirfmt: SingleLanePerSampleSingleEndFastqDirFmt) -> _PlotQualView:
-    return _PlotQualView(dirfmt, paired=False)
-
-
-@plugin.register_transformer
-def _31(dirfmt: SingleLanePerSamplePairedEndFastqDirFmt) -> _PlotQualView:
-    return _PlotQualView(dirfmt, paired=True)
-
-
 @plugin.register_transformer
 def _32(data: pd.DataFrame) -> ErrorCorrectionDetailsFmt:
     ff = ErrorCorrectionDetailsFmt()
@@ -105,3 +102,54 @@ def _33(ff: ErrorCorrectionDetailsFmt) -> pd.DataFrame:
 @plugin.register_transformer
 def _34(ff: ErrorCorrectionDetailsFmt) -> Metadata:
     return Metadata.load(str(ff))
+
+
+@plugin.register_transformer
+def _65(dirfmt: EMPSingleEndDirFmt) -> BarcodeSequenceFastqIterator:
+    barcode_generator = _read_fastq_seqs(
+        str(dirfmt.barcodes.view(FastqGzFormat)))
+    sequence_generator = _read_fastq_seqs(
+        str(dirfmt.sequences.view(FastqGzFormat)))
+    result = BarcodeSequenceFastqIterator(barcode_generator,
+                                          sequence_generator)
+    # ensure that dirfmt stays in scope as long as result does so these
+    # generators will work.
+    result.__dirfmt = dirfmt
+    return result
+
+
+# TODO: remove this when names are aliased
+@plugin.register_transformer
+def _65_legacy(dirfmt: EMPMultiplexedDirFmt) -> BarcodeSequenceFastqIterator:
+    return _65(dirfmt)
+
+
+@plugin.register_transformer
+def _67(dirfmt: EMPPairedEndDirFmt) -> BarcodeSequenceFastqIterator:
+    barcode_generator = _read_fastq_seqs(
+        str(dirfmt.barcodes.view(FastqGzFormat)))
+    sequence_generator = _read_fastq_seqs(
+        str(dirfmt.forward.view(FastqGzFormat)))
+    result = BarcodeSequenceFastqIterator(barcode_generator,
+                                          sequence_generator)
+    # ensure that dirfmt stays in scope as long as result does so these
+    # generators will work.
+    result.__dirfmt = dirfmt
+    return result
+
+
+@plugin.register_transformer
+def _68(dirfmt: EMPPairedEndDirFmt) -> BarcodePairedSequenceFastqIterator:
+    barcode_generator = _read_fastq_seqs(
+        str(dirfmt.barcodes.view(FastqGzFormat)))
+    forward_generator = _read_fastq_seqs(
+        str(dirfmt.forward.view(FastqGzFormat)))
+    reverse_generator = _read_fastq_seqs(
+        str(dirfmt.reverse.view(FastqGzFormat)))
+    result = BarcodePairedSequenceFastqIterator(barcode_generator,
+                                                forward_generator,
+                                                reverse_generator)
+    # ensure that dirfmt stays in scope as long as result does so these
+    # generators will work.
+    result.__dirfmt = dirfmt
+    return result
