+++ q2_types/multiplexed_sequences/tests/test_transformer.py
@@ -9,6 +9,7 @@
 import unittest
 import os
 import string
+import tempfile
 
 import pandas as pd
 import pandas.testing as pdt
@@ -16,13 +17,17 @@ import skbio
 
 import qiime2
 from qiime2.plugin.testing import TestPluginBase
+from qiime2.plugin import ValidationError
 
 from q2_types.multiplexed_sequences import (
     MultiplexedFastaQualDirFmt,
     MultiplexedSingleEndBarcodeInSequenceDirFmt,
-    ErrorCorrectionDetailsFmt
+    ErrorCorrectionDetailsFmt,
+    EMPSingleEndDirFmt, EMPSingleEndCasavaDirFmt,
+    BarcodeSequenceFastqIterator
 
 )
+
 from q2_types.per_sample_sequences import (
     SingleLanePerSampleSingleEndFastqDirFmt,
     SingleLanePerSamplePairedEndFastqDirFmt,
@@ -218,6 +223,247 @@ class TestErrorCorrectionDetailsFmtTransformers(TestPluginBase):
         obs = transformer(ff)
 
         self.assertEqual(obs, qiime2.Metadata(self.df))
+
+
+class TestTransformers(TestPluginBase):
+    package = 'q2_types.feature_data.tests'
+
+    def setUp(self):
+        # TODO generalize plugin lookup when ported to framework. This code
+        # is adapted from the base class.
+        try:
+            from q2_types.plugin_setup import plugin
+        except ImportError:
+            self.fail("Could not import plugin object.")
+
+        self.plugin = plugin
+
+        # TODO use qiime temp dir when ported to framework, and when the
+        # configurable temp dir exists
+        self.temp_dir = tempfile.TemporaryDirectory(
+            prefix='q2-types-test-temp-')
+
+    def test_emp_multiplexed_format_barcode_sequence_iterator(self):
+        transformer = self.get_transformer(EMPSingleEndDirFmt,
+                                           BarcodeSequenceFastqIterator)
+        dirname = 'emp_multiplexed'
+        dirpath = self.get_data_path(dirname)
+        bsi = transformer(EMPSingleEndDirFmt(dirpath, mode='r'))
+        bsi = list(bsi)
+        self.assertEqual(len(bsi), 250)
+        self.assertEqual(
+            bsi[0][0],
+            ('@M00176:17:000000000-A0CNA:1:1:15487:1773 1:N:0:0',
+             'TTAGGCATCTCG',
+             '+',
+             'B@@FFFFFHHHH'))
+        self.assertEqual(
+            bsi[0][1],
+            ('@M00176:17:000000000-A0CNA:1:1:15487:1773 1:N:0:0',
+             'GCTTAGGGATTTTATTGTTATCAGGGTTAATCGTGCCAAGAAAAGCGGCATGGTCAATATAAC'
+             'CAGTAGTGTTAACAGTCGGGAGAGGAGTGGCATTAACACCATCCTTCATGAACTTAATCCACT'
+             'GTTCACCATAAACGTGACGATGAGG',
+             '+',
+             'C@CFFFFFHHFHHGIJJ?FFHEIIIIHGEIIFHGIIJHGIGBGB?DHIIJJJJCFCHIEGIGG'
+             'HGFAEDCEDBCCEEA.;>?BB=288A?AB709@:3:A:C88CCD@CC444@>>34>>ACC:?C'
+             'CD<CDCA>A@A>:<?B@?<((2(>?'))
+
+    def test_emp_se_multiplexed_format_barcode_sequence_iterator(self):
+        transformer1 = self.get_transformer(EMPSingleEndCasavaDirFmt,
+                                            EMPSingleEndDirFmt)
+        transformer2 = self.get_transformer(EMPSingleEndDirFmt,
+                                            BarcodeSequenceFastqIterator)
+        dirname = 'emp_multiplexed_single_end'
+        dirpath = self.get_data_path(dirname)
+        emp_demultiplexed = \
+            transformer1(EMPSingleEndCasavaDirFmt(dirpath, mode='r'))
+        bsi = transformer2(EMPSingleEndDirFmt(emp_demultiplexed, mode='r'))
+        bsi = list(bsi)
+        self.assertEqual(len(bsi), 250)
+        self.assertEqual(
+            bsi[0][0],
+            ('@M00176:17:000000000-A0CNA:1:1:15487:1773 1:N:0:0',
+             'TTAGGCATCTCG',
+             '+',
+             'B@@FFFFFHHHH'))
+        self.assertEqual(
+            bsi[0][1],
+            ('@M00176:17:000000000-A0CNA:1:1:15487:1773 1:N:0:0',
+             'GCTTAGGGATTTTATTGTTATCAGGGTTAATCGTGCCAAGAAAAGCGGCATGGTCAATATAAC'
+             'CAGTAGTGTTAACAGTCGGGAGAGGAGTGGCATTAACACCATCCTTCATGAACTTAATCCACT'
+             'GTTCACCATAAACGTGACGATGAGG',
+             '+',
+             'C@CFFFFFHHFHHGIJJ?FFHEIIIIHGEIIFHGIIJHGIGBGB?DHIIJJJJCFCHIEGIGG'
+             'HGFAEDCEDBCCEEA.;>?BB=288A?AB709@:3:A:C88CCD@CC444@>>34>>ACC:?C'
+             'CD<CDCA>A@A>:<?B@?<((2(>?'))
+
+    def test_invalid(self):
+        dirname = 'bad'
+        dirpath = self.get_data_path(dirname)
+        transformer = self.get_transformer(EMPSingleEndDirFmt,
+                                           BarcodeSequenceFastqIterator)
+        with self.assertRaises(ValidationError):
+            transformer(EMPSingleEndDirFmt(dirpath, mode='r'))
+
+        transformer = self.get_transformer(EMPSingleEndCasavaDirFmt,
+                                           EMPSingleEndDirFmt)
+        with self.assertRaises(ValidationError):
+            transformer(EMPSingleEndCasavaDirFmt(dirpath, 'r'))
+
+
+class BarcodeSequenceFastqIteratorTests(unittest.TestCase):
+
+    def test_valid(self):
+        barcodes = [('@s1/2 abc/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2 abc/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1 abc/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1 abc/1', 'AAA', '+', 'PPP'),
+                     ('@s4/1 abc/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        for i, (barcode, sequence) in enumerate(bsi):
+            self.assertEqual(barcode, barcodes[i])
+            self.assertEqual(sequence, sequences[i])
+
+    def test_too_few_barcodes(self):
+        barcodes = [('@s1/2 abc/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1 abc/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1 abc/1', 'AAA', '+', 'PPP'),
+                     ('@s4/1 abc/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+    def test_too_few_sequences(self):
+        barcodes = [('@s1/2 abc/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2 abc/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc/1', 'GGG', '+', 'YYY')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+    def test_mismatched_id(self):
+        barcodes = [('@s1/2 abc/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2 abc/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1 abc/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1 abc/1', 'AAA', '+', 'PPP'),
+                     ('@s5/1 abc/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+    def test_mismatched_description(self):
+        barcodes = [('@s1/2 abc/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2 abc/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1 abc/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1 abc/1', 'AAA', '+', 'PPP'),
+                     ('@s4/1 abd/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+    def test_mismatch_description_override(self):
+        barcodes = [('@s1/2 abc/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2 abc/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1 abc/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1 abc/1', 'AAA', '+', 'PPP'),
+                     ('@s4/1 abd/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences,
+                                           ignore_description_mismatch=True)
+        self.assertEqual(len(list(bsi)), 4)
+
+    def test_mismatched_handles_slashes_in_id(self):
+        # mismatch is detected as being before the last slash, even if there
+        # is more than one slash
+        barcodes = [('@s1/2/2 abc/2', 'AAAA', '+', 'YYYY')]
+        sequences = [('@s1/1/1 abc/1', 'GGG', '+', 'YYY')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+    def test_mismatched_handles_slashes_in_description(self):
+        # mismatch is detected as being before the last slash, even if there
+        # is more than one slash
+        barcodes = [('@s1/2 a/2/2', 'AAAA', '+', 'YYYY')]
+        sequences = [('@s1/1 a/1/1', 'GGG', '+', 'YYY')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+    def test_no_description(self):
+        barcodes = [('@s1/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1', 'AAA', '+', 'PPP'),
+                     ('@s4/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        for i, (barcode, sequence) in enumerate(bsi):
+            self.assertEqual(barcode, barcodes[i])
+            self.assertEqual(sequence, sequences[i])
+
+    def test_only_one_description(self):
+        barcodes = [('@s1/2 abc', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2 abc', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2 abc', 'AACC', '+', 'PPPP'),
+                    ('@s4/2 abc', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1', 'GGG', '+', 'YYY'),
+                     ('@s2/1', 'CCC', '+', 'PPP'),
+                     ('@s3/1', 'AAA', '+', 'PPP'),
+                     ('@s4/1', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
+
+        barcodes = [('@s1/2', 'AAAA', '+', 'YYYY'),
+                    ('@s2/2', 'AAAA', '+', 'PPPP'),
+                    ('@s3/2', 'AACC', '+', 'PPPP'),
+                    ('@s4/2', 'AACC', '+', 'PPPP')]
+
+        sequences = [('@s1/1 abc', 'GGG', '+', 'YYY'),
+                     ('@s2/1 abc', 'CCC', '+', 'PPP'),
+                     ('@s3/1 abc', 'AAA', '+', 'PPP'),
+                     ('@s4/1 abc', 'TTT', '+', 'PPP')]
+
+        bsi = BarcodeSequenceFastqIterator(barcodes, sequences)
+        with self.assertRaises(ValueError):
+            list(bsi)
 
 
 if __name__ == '__main__':
