+++ q2_types/plugin_setup.py
@@ -11,7 +11,6 @@ import importlib
 import pandas as pd
 
 import qiime2.plugin
-import qiime2.sdk
 
 from q2_types import __version__
 
@@ -31,3 +30,4 @@ plugin.register_views(pd.Series, pd.DataFrame,
 
 importlib.import_module('q2_types.bowtie2._deferred_setup')
 importlib.import_module('q2_types.distance_matrix._deferred_setup')
+importlib.import_module('q2_types.feature_data._deferred_setup')
